src/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
google-generativeai>=0.3.2
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
pymongo>=4.9.0,<5.0.0
motor>=3.3.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
email-validator>=2.1.0

# Dependências de teste
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.24.0